refactor: move inbox collection functions inside inbox collection file, don't export collection
